@@ -1,18 +1,22 @@
 {
   "name": "reactjs-socket-io-chat-app",
   "version": "0.6.9",
+  "proxy": "http://localhost:3001/",
   "private": true,
   "scripts": {
-    "build": "browserify ./client/app.jsx -t babelify --outfile ./public/js/app.js",
-    "start": "node app.js"
-  },
-  "dependencies": {
-    "express": "^4.13.1",
-    "react": "^0.13.3",
-    "socket.io": "^1.3.5"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "babelify": "^6.1.3",
-    "browserify": "^10.2.6"
+    "babel-cli": "^6.24.1",
+    "react-scripts": "1.0.7"
+  },
+  "dependencies": {
+    "express": "^4.15.3",
+    "react": "^15.6.1",
+    "socket.io": "^1.7.4",
+    "react-dom": "^15.6.1"
   }
 }
